Update Laminas library and make improvements to test suite
The Laminas library has been updated to the latest version, which includes various dependency updates. Additionally, improvements have been made to the test suite by adding more attributes to classes for better readability and maintainability. The new attributes will help track the state and behavior of instances throughout tests. The tests will be more independent and easier to understand as a result.

Signed-off-by: Sebastian Hopfe <s.hopfe@me.com>